Updated with recommendations from Copilot
