+++ .gitignore
@@ -32,5 +32,6 @@
 *.app
 
 # Folders
+.idea/
 .vscode/
 .venv/
